web-platform_23_06_dev/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
black==23.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cx-Oracle==8.3.0
Django==4.1.7
django-cors-headers==3.14.0
django-environ==0.10.0
django-grappelli==3.0.4
djangorestframework==3.14.0
et-xmlfile==1.1.0
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
openpyxl==3.1.2
packaging==23.0
pathspec==0.11.1
Pillow==9.4.0
platformdirs==3.2.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
yarl==1.8.2
